AMR-WB: the bit-exact decoder is reachable from outside the crate
`wb::decoder::Decoder` has matched TS 26.173 on every sample of every rate
for several commits, and none of it was usable: `AmrCodec::decode` still
returned `FeatureNotEnabled`, so nothing outside this crate could decode AMR.
That gap is closed.

`AudioCodec::decode` takes bare bytes with nowhere to say what mode they are
in, which for AMR would normally be fatal — the mode is an input, not
something the payload states. It works here because within one variant every
speech mode has a distinct octet-aligned length, so the length identifies the
mode. That is a bijection rather than a heuristic and there is a test
asserting it stays one; it does not hold across variants, since narrowband
12.2 and wideband 8.85 are both 31 bytes, which is why the variant is fixed
at construction. A length matching no mode is an error rather than a guess.

`VariableRateCodec::decode_frame` is the interface the RTP path should use:
it carries the mode, the RFC 4867 Q bit and the frame kind explicitly. A
damaged frame is refused rather than decoded as if it were good, because
decoding damaged bits produces a loud artefact and concealment — the thing
that exists to prevent exactly that — is not implemented yet. SID, NO_DATA
and lost frames say which missing piece each needs.

The test that matters decodes the committed reference fixtures through the
*public API* at all nine rates and compares against TS 26.173's own output,
masked to the 14 bits AMR-WB defines. Per-stage exactness cannot see a
wiring layer that resets state between frames or loses a frame boundary;
this can, and there is a companion test asserting that `reset` really does
return the decoder to its start state rather than leaving one call's tail in
the next.

Narrowband decoding and both encoders still refuse, naming what they need.

+++ crates/media/codec-core/src/codecs/amr/nb/math.rs
@@ -200,9 +200,9 @@ mod tests {
     #[test]
     fn log2_recovers_the_exponent_of_a_power_of_two() {
         let mut c = ctx();
-        for shift in 1..=30 {
+        for shift in 1..=30i16 {
             let (exponent, fraction) = log2(&mut c, Word32(1 << shift));
-            assert_eq!(exponent.0, shift as i16, "log2(2^{shift})");
+            assert_eq!(exponent.0, shift, "log2(2^{shift})");
             // A power of two lands exactly on a table entry, so the fraction is
             // zero rather than merely small. An index base off by one would
             // show up here and nowhere else in this test.
